+++ Player.cs
@@ -49,7 +49,7 @@ namespace RPG
         }
 
         //
-        public override void takeDamage(int d)
+        public override void takeDamage(double d)
         {
             if (random.Next(5) != 1)
             {
@@ -65,7 +65,7 @@ namespace RPG
 
             else
             {
-                currHP -= (int)(d * DEF);
+                currHP -= (d * DEF);
 
                 Console.WriteLine(
